requirements_py2.txt
numpy>=1.15.1
onnx==1.4.1
Pillow>=5.2.0
wget>=3.2
click
